feat: 소비 패턴 분석 기능 일단은 돌아가게 완성

+++ backend/LocketCommon/src/main/java/com/locket/elastic/dto/FeedbackAgeGroupComparisonDto.java
@@ -4,12 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class FeedbackAgeGroupComparisonDto {
-    private Map<String, Integer> userSpending;
+    private Map<String, Double> userSpending;
     private Map<String, Double> ageGroupAverage;
 }
