+++ backend/LocketCommon/src/main/java/com/locket/elastic/dto/FeedbackMonthlyChangeDto.java
@@ -4,14 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class FeedbackMonthlyChangeDto {
-    private int currentMonthTotal;
-    private int previousMonthTotal;
-    private double totalChangeRate;
+    private Double currentMonthTotal;
+    private Double previousMonthTotal;
+    private Double totalChangeRate;
     private Map<String, Double> categoryChangeRate;
 }
